requirements.txt
acres==0.5.0
aiofiles==24.1.0
altair==5.5.0
altgraph==0.17.4
anyio==4.9.0
attrs==25.3.0
auto-py-to-exe==2.46.0
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
Eel==0.18.1
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
frontend==0.0.3
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.44

greenlet==3.2.3
gspread==6.2.1
h11==0.16.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.44.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.0
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
pathlib==1.0.1
patsy==1.0.1
pdf2image==1.17.0
pefile==2023.2.7
pillow==10.4.0
plotly==6.2.0
plotly-express==0.4.1
protobuf==5.29.5
prov==2.0.2
puremagic==1.29
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
qrcode==8.2
rdflib==7.1.4
referencing==0.36.2
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.0
statsmodels==0.14.4
streamlit==1.36.0
streamlit-aggrid==1.1.5.post1



tenacity==8.5.0
toml==0.10.2
tools==1.0.2
tornado==6.5.1
traits==7.0.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
watchdog==4.0.2
xlsxwriter==3.2.5
zope.event==5.0
zope.interface==7.2

streamlit-option-menu==0.4.0
pyzbar
qrcode
opencv-python
streamlit-webrtc
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.149.0